Move other file to decorators in shogi.py
